Added new price tier

+++ backend/Api.Abstractions/ServerTier.cs
@@ -3,7 +3,7 @@ namespace Api.Abstractions;
 public sealed record ServerTier(
     ServerTierId Id,
     decimal RamGb,
-    int Cores,
+    decimal Cores,
     decimal DiskGb,
     Price Price,
     string? PromotionalText
